EHN add parameter axis in safe_indexing to slice rows and columns
